Printint a uint8_t doesn't do what I expected.

+++ programs/program3/src/common.cpp
@@ -23,7 +23,7 @@ void PrintSolution( std::vector<uint8_t> solution )
 {
     for( auto&& q : solution )
     {
-        std::cout << q << " ";
+        std::cout << static_cast<int>( q ) << " ";
     }
     std::cout << std::endl;
 }
